requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confluent-kafka==2.1.1
Deprecated==1.2.14
Flask==2.3.2
googleapis-common-protos==1.52.0
grpcio==1.54.2
grpcio-tools==1.54.2
idna==3.4
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
opentelemetry-api==1.18.0
opentelemetry-exporter-jaeger==1.18.0
opentelemetry-exporter-jaeger-proto-grpc==1.18.0
opentelemetry-exporter-jaeger-thrift==1.18.0
opentelemetry-propagator-b3==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
protobuf==4.23.2
requests==2.31.0
six==1.16.0
thrift==0.16.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.4
wrapt==1.15.0
zipp==3.15.0
